backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
pydantic-settings
pydantic[email]
slowapi
asyncpg
supabase
httpx
python-jose[cryptography]
cachetools
certifi
python-multipart
alembic
python-dotenv
email-validator
pdfplumber
ezdxf
shapely
pillow
